# New option -noast to generate parser w/out AST support