--- a/peg.go
+++ b/peg.go
@@ -145,7 +145,9 @@ type {{.StructName}} struct {
 	parse		func(rule ...int) error
 	reset		func()
 	Pretty 	bool
-{{if .Ast}}    tokens32{{end}}
+{{if .Ast -}}
+	tokens32
+{{- end}}
 }
 
 func (p *{{.StructName}}) Parse(rule ...int) error {
@@ -237,6 +239,9 @@ func (p *{{.StructName}}) Init() {
 		max token32
 		position, tokenIndex uint32
 		buffer []rune
+{{if not .Ast -}}
+		text string
+{{- end}}
 	)
 	p.reset = func() {
 		max = token32{}
@@ -250,24 +255,32 @@ func (p *{{.StructName}}) Init() {
 	}
 	p.reset()
 
-    _rules := p.rules
-{{if .Ast}} tree := tokens32{tree: make([]token32, math.MaxInt16)}{{end}}
+	_rules := p.rules
+{{if .Ast -}}
+	tree := tokens32{tree: make([]token32, math.MaxInt16)}
+{{- end}}
 	p.parse = func(rule ...int) error {
 		r := 1
 		if len(rule) > 0 {
 			r = rule[0]
 		}
 		matches := p.rules[r]()
-{{if .Ast}}     p.tokens32 = tree{{end}}
+{{if .Ast -}}
+		p.tokens32 = tree
+{{- end}}
 		if matches {
-{{if .Ast}}         p.Trim(tokenIndex){{end}}
+{{if .Ast -}}
+			p.Trim(tokenIndex)
+{{- end}}
 			return nil
 		}
 		return &parseError{p, max}
 	}
 
 	add := func(rule pegRule, begin uint32) {
-{{if .Ast}} tree.Add(rule, begin, position, tokenIndex){{end}}
+{{if .Ast -}}
+		tree.Add(rule, begin, position, tokenIndex)
+{{- end}}
 		tokenIndex++
 		if begin != position && position > max.end {
 			max = token32{rule, begin, position}
@@ -1206,11 +1219,24 @@ func (t *Tree) Compile(file string, out io.Writer) {
 			nodeType, rule := element.GetType(), element.Next()
 			printBegin()
 			if nodeType == TypeAction {
-				_print("\nadd(rule%v, position)", rule)
+				if t.Ast {
+					_print("\nadd(rule%v, position)", rule)
+				} else {
+					// There is no AST support, so inline the rule code
+					_print("\n%v", element)
+				}
 			} else {
 				_print("\nposition%d := position", ok)
 				compile(element, ko)
-				_print("\nadd(rule%v, position%d)", rule, ok)
+				if n.GetType() == TypePush && !t.Ast {
+					// This is TypePush and there is no AST support,
+					// so inline capture to text right here
+					_print("\nbegin := position%d", ok)
+					_print("\nend := position")
+					_print("\ntext = string(buffer[begin:end])")
+				} else {
+					_print("\nadd(rule%v, position%d)", rule, ok)
+				}
 			}
 			printEnd()
 		case TypeAlternate: