Implements #53 and improve #51.

Didn`t dare to check if peg itself could be made with such option. 

As per #51 benchcmp for extremely stupid grammar with actions (see at https://github.com/egorse/peg_ex)
```
benchmark           old ns/op     new ns/op     delta
Benchmark100-4      98626         6583          -93.33%
Benchmark200-4      106745        11241         -89.47%
Benchmark500-4      125213        25230         -79.85%
Benchmark1k-4       154281        50276         -67.41%
Benchmark10k-4      689338        482092        -30.06%
Benchmark20k-4      678916        483708        -28.75%
Benchmark50k-4      3255189       2337557       -28.19%
Benchmark100k-4     6744704       4654680       -30.99%
Benchmark200k-4     14254404      9987951       -29.93%
Benchmark500k-4     36308161      26848004      -26.06%
Benchmark1M-4       83490497      55150098      -33.94%

benchmark           old allocs     new allocs     delta
Benchmark100-4      61             60             -1.64%
Benchmark200-4      104            103            -0.96%
Benchmark500-4      231            230            -0.43%
Benchmark1k-4       456            455            -0.22%
Benchmark10k-4      4254           4253           -0.02%
Benchmark20k-4      4254           4253           -0.02%
Benchmark50k-4      21115          21113          -0.01%
Benchmark100k-4     42202          42199          -0.01%
Benchmark200k-4     84374          84370          -0.00%
Benchmark500k-4     210870         210864         -0.00%
Benchmark1M-4       431823         431816         -0.00%

benchmark           old bytes     new bytes     delta
Benchmark100-4      394752        1472          -99.63%
Benchmark200-4      395680        2400          -99.39%
Benchmark500-4      397920        4640          -98.83%
Benchmark1k-4       402528        9248          -97.70%
Benchmark10k-4      495136        101856        -79.43%
Benchmark20k-4      495136        101856        -79.43%
Benchmark50k-4      1669664       489952        -70.66%
Benchmark100k-4     3715745       963168        -74.08%
Benchmark200k-4     7938976       2040672       -74.30%
Benchmark500k-4     30062176      5289506       -82.40%
Benchmark1M-4       61947872      12009376      -80.61%
```