@tj 2017-01-16
Should this break parsers? I'm building my own AST in the actions, but -noast seems to break their semantics, yielding a different custom AST in my case.