@pointlander 2017-01-16
-noast can break your parser if there is backtracking with you actions

On Mon, Jan 16, 2017 at 1:41 PM, TJ Holowaychuk <notifications@github.com>
wrote:

> Should this break parsers? I'm building my own AST in the actions, but
> -noast seems to break their semantics, yielding a different custom AST in
> my case.
>
> —
> You are receiving this because you modified the open/close state.
> Reply to this email directly, view it on GitHub
> <https://github.com/pointlander/peg/pull/60#issuecomment-272957531>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAPZFJWJSAImQvuKFldPozGu5fC55jK_ks5rS9XugaJpZM4LRPjZ>
> .
>